Move el_geometry into utils/_clique
