+++ minorminer/utils/_clique/el_geometry.py
