+++ setup.py
@@ -141,11 +141,9 @@ extensions = [
 setup(
     ext_modules=cythonize(extensions),
     packages=['minorminer',
-              'minorminer.embedders',
-              'minorminer.embedders.clique',
-              'minorminer.embedders.clique._zephyr',
               'minorminer.layout',
               'minorminer.utils',
+              'minorminer.utils._clique',
               'minorminer._extern',
               'minorminer._extern.rpack',
               ],
